CRUD ETUDIANT

+++ back/src/main/java/com/example/springproject/Repository/EtudiantRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EtudiantRepository extends CrudRepository<Etudiant,Long> {
+public interface EtudiantRepository extends CrudRepository<Etudiant,Integer> {
 }
